requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
executing==2.2.0
filelock==3.13.1
filetype==1.2.0
frozenlist==1.6.2
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
griffe==1.7.3
gTTS==2.5.4
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
llama-cloud==0.1.23
llama-cloud-services==0.6.30
llama-index==0.12.40
llama-index-agent-openai==0.4.9
llama-index-cli==0.4.3
llama-index-core==0.12.40
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.4
llama-index-llms-openai==0.4.3
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.30
llvmlite==0.44.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.41.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
openai==1.84.0
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
playsound==1.2.2
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycairo==1.28.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyGObject==3.52.3
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
stack-data==0.6.3
streamlit==1.45.0
striprtf==0.0.26
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.3.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.20.0
